requirements.txt
streamlit==1.20.0
sentence-transformers==2.2.2
pinecone-client==2.2.1
uuid
openai==0.27.2
langchain==0.0.281
auto-gptq[llama]
transformers
spacy
PyPDF2




